Reports works 100%, the discount concept is works, the users and the complains and the orders is saved in the server side.

+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/DeliveryPageController.java
@@ -256,7 +256,7 @@ public class DeliveryPageController {
             addressField.setVisible(true);
             homeNumberField.setVisible(true);
             done_Order.setOrderType("Delivery");
-            int newPrice = done_Order.getTotal_price() + deliveryPrice;
+            double newPrice = done_Order.getTotal_price() + deliveryPrice;
             price.setText(newPrice + "₪");
         }
     }
