+++ entities/src/main/java/il/cshaifasweng/OCSFMediatorExample/entities/Order.java
@@ -21,7 +21,7 @@ public class Order implements Serializable{
     private String restaurantName;
     private LocalDate date;
     private String orderType;
-    private int total_price;
+    private double total_price;
     private String orderStatus;
     private String customerEmail;
     private LocalDateTime orderTime;
@@ -98,11 +98,11 @@ public class Order implements Serializable{
         this.orderType = orderType;
     }
 
-    public int getTotal_price() {
+    public double getTotal_price() {
         return total_price;
     }
 
-    public void setTotal_price(int total_price) {
+    public void setTotal_price(double total_price) {
         this.total_price = total_price;
     }
